options-pricing-engine/pyproject.toml
[project]
name = "options-pricing-engine"
version = "1.0.1"
description = "Production-grade real-time options pricing engine"
authors = [{name = "UCTG/1 Ultimate Coding Organization"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "pydantic>=2.6.0,<3",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "prometheus-client>=0.19.0",
    "psutil>=5.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "redis>=5.0.0"
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0","pytest-cov>=4.1.0","black>=24.3.0","httpx>=0.26.0"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']
